LPD-10653 LPD-53772 search-experiences-service: move logic up hierarchy
https://liferay.atlassian.net/browse/LPD-53772
