+++ modules/dxp/apps/search-experiences/search-experiences-service/src/main/java/com/liferay/search/experiences/internal/info/collection/provider/BlogsEntrySXPBlueprintInfoCollectionProvider.java
@@ -12,17 +12,13 @@ import com.liferay.info.collection.provider.CollectionQuery;
 import com.liferay.info.collection.provider.FilteredInfoCollectionProvider;
 import com.liferay.info.collection.provider.SingleFormVariationInfoCollectionProvider;
 import com.liferay.info.pagination.InfoPage;
-import com.liferay.info.pagination.Pagination;
 import com.liferay.petra.function.transform.TransformUtil;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.search.Field;
 import com.liferay.portal.search.document.Document;
 import com.liferay.portal.search.hits.SearchHits;
-import com.liferay.portal.search.searcher.SearchRequestBuilder;
 import com.liferay.portal.search.searcher.SearchRequestBuilderFactory;
 import com.liferay.portal.search.searcher.SearchResponse;
 import com.liferay.portal.search.searcher.Searcher;
@@ -30,7 +26,6 @@ import com.liferay.search.experiences.model.SXPBlueprint;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Locale;
 
 /**
  * @author Joshua Cords
@@ -56,43 +51,19 @@ public class BlogsEntrySXPBlueprintInfoCollectionProvider
 		CollectionQuery collectionQuery) {
 
 		try {
-			Pagination pagination = collectionQuery.getPagination();
-
-			SearchRequestBuilder searchRequestBuilder = getSearchRequestBuilder(
-				collectionQuery, pagination);
-
-			SearchResponse searchResponse = searcher.search(
-				searchRequestBuilder.build());
+			SearchResponse searchResponse = getCollectionQuerySearchResponse(
+				collectionQuery);
 
 			return InfoPage.of(
 				_getBlogEntries(searchResponse.getSearchHits()),
 				collectionQuery.getPagination(), searchResponse.getTotalHits());
 		}
 		catch (Exception exception) {
-			_log.error("Unable to get journal articles", exception);
+			_log.error("Unable to get blog entries", exception);
 		}
 
 		return InfoPage.of(
 			Collections.emptyList(), collectionQuery.getPagination(), 0);
-	}
-
-	@Override
-	public String getFormVariationKey() {
-		return sxpBlueprint.getExternalReferenceCode();
-	}
-
-	@Override
-	public String getKey() {
-		return StringBundler.concat(
-			BlogsEntrySXPBlueprintInfoCollectionProvider.class.getName(),
-			StringPool.UNDERLINE, sxpBlueprint.getCompanyId(),
-			StringPool.UNDERLINE, sxpBlueprint.getExternalReferenceCode(),
-			StringPool.UNDERLINE, BlogsEntry.class.getName());
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		return sxpBlueprint.getTitle(locale);
 	}
 
 	private List<BlogsEntry> _getBlogEntries(SearchHits searchHits)
