+++ modules/dxp/apps/search-experiences/search-experiences-service/src/main/java/com/liferay/search/experiences/internal/info/collection/provider/SXPBlueprintInfoCollectionProvider.java
@@ -15,6 +15,8 @@ import com.liferay.info.filter.InfoFilter;
 import com.liferay.info.filter.KeywordsInfoFilter;
 import com.liferay.info.filter.TagsInfoFilter;
 import com.liferay.info.pagination.Pagination;
+import com.liferay.petra.string.StringBundler;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
@@ -24,6 +26,7 @@ import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.search.constants.SearchContextAttributes;
 import com.liferay.portal.search.searcher.SearchRequestBuilder;
 import com.liferay.portal.search.searcher.SearchRequestBuilderFactory;
+import com.liferay.portal.search.searcher.SearchResponse;
 import com.liferay.portal.search.searcher.Searcher;
 import com.liferay.search.experiences.model.SXPBlueprint;
 
@@ -47,6 +50,30 @@ public abstract class SXPBlueprintInfoCollectionProvider<T>
 		this.searcher = searcher;
 		this.searchRequestBuilderFactory = searchRequestBuilderFactory;
 		this.sxpBlueprint = sxpBlueprint;
+	}
+
+	public SearchResponse getCollectionQuerySearchResponse(
+		CollectionQuery collectionQuery) {
+
+		Pagination pagination = collectionQuery.getPagination();
+
+		SearchRequestBuilder searchRequestBuilder = getSearchRequestBuilder(
+			collectionQuery, pagination);
+
+		return searcher.search(searchRequestBuilder.build());
+	}
+
+	@Override
+	public String getFormVariationKey() {
+		return sxpBlueprint.getExternalReferenceCode();
+	}
+
+	@Override
+	public String getKey() {
+		return StringBundler.concat(
+			SXPBlueprint.class.getName(), StringPool.UNDERLINE,
+			sxpBlueprint.getCompanyId(), StringPool.UNDERLINE,
+			sxpBlueprint.getExternalReferenceCode());
 	}
 
 	@Override
